Rename tokens module to commands
