+++ src/lib.rs
@@ -1,5 +1,5 @@
 pub mod arg_parser;
 pub mod args;
+pub mod commands;
 mod db;
 pub mod thin_client;
-pub mod tokens;
